requirements.txt
requests~=2.31.0
pandas~=2.2.0
beautifulsoup4~=4.12.3
google~=3.0.0
google-auth==2.28.1
google-api-python-client==2.121.0
python-dotenv==1.0.1